requirements-dev.txt
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1
python-dotenv==1.0.0
httpx==0.24.1
pytest-cov==4.1.0
